requirements.txt
uv
streamlit
langchain
langchain_groq
langchain_community
wikipedia
python-dotenv
numexpr
